Move type to types.
